# api/openapi.yaml
openapi: 3.0.3
info:
  title: Hydra API
  version: 1.0.0
paths:
...
  /api/projects/{project_id}/agents/{id}/diff:
    get:
      summary: Get the diff for an agent's branch
      operationId: getAgentDiff
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
          description: Project ID
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: base_ref
          required: false
          schema:
            type: string
          description: Base commit SHA or ref. Defaults to the agent's base branch.
        - in: query
          name: head_ref
          required: false
          schema:
            type: string
          description: Head commit SHA or ref. Defaults to the agent's branch.
        - in: query
          name: ignore_whitespace
          required: false
          schema:
            type: boolean
          description: Ignore whitespace changes in the diff
        - in: query
          name: include_uncommitted
          required: false
          schema:
            type: boolean
          description: Include uncommitted changes in the worktree in the diff
        - in: query
          name: path
          required: false
          schema:
            type: string
          description: Only return the diff for this specific file path
        - in: query
          name: context
          required: false
          schema:
            type: integer
            default: 3
          description: Number of lines of context to show (defaults to 3)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiffResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/projects/{project_id}/agents/{id}/artifacts:
    get:
      summary: Get generated visual artifacts (e.g. screenshots) for both sides of a diff
      description: >
        Returns, per configured artifact script, the generated image files for
        the left and right versions of the comparison and whether they differ.
        Generation runs in the background and is cached; a script with status
        "generating" should be polled. Returns an empty list when the project
        configures no artifact scripts.
      operationId: getAgentArtifacts
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
          description: Project ID
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: base_ref
          required: false
          schema:
            type: string
          description: Left (base) commit SHA or ref. Defaults to the agent's base branch.
        - in: query
          name: head_ref
          required: false
          schema:
            type: string
          description: Right (head) commit SHA or ref. Defaults to the agent's branch tip.
        - in: query
          name: include_uncommitted
          required: false
          schema:
            type: boolean
          description: Use the agent's uncommitted working tree as the right version.
        - in: query
          name: refresh
          required: false
          schema:
            type: string
          description: >
            Name of a single artifact script whose cached result (including a
            cached failure) should be discarded and regenerated for both sides
            of the comparison before responding.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactsResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      required: [error, code, details]
      properties:
        error:
          type: string
          description: "Machine-readable error type (e.g. internal_error, not_found, unauthorized, docker_connect)"
          enum: [not_found, unauthorized, internal_error, docker_connect, path_not_found, not_a_git_repo, bad_request, conflict]
        code:
          type: integer
          description: "HTTP status code"
        details:
          type: string
          description: "Human-readable error description"
...
    CommitInfo:
      type: object
      required: [sha, short_sha, message, author_name, author_email, timestamp]
      properties:
        sha:
          type: string
          description: Full 40-character commit SHA
        short_sha:
          type: string
          description: Abbreviated 7-character commit SHA
        message:
          type: string
          description: Full commit message
        subject:
          type: string
          description: First line of the commit message
        author_name:
          type: string
        author_email:
          type: string
        timestamp:
          type: string
          description: ISO 8601 timestamp of the author date
...
    DiffLine:
      type: object
      required: [type, content]
      properties:
        type:
          type: string
          enum: [context, addition, deletion, no_newline]
          description: "context: unchanged line; addition: added line; deletion: removed line; no_newline: no newline at end of file"
        content:
          type: string
          description: Line content without the diff prefix character
        old_line_num:
          type: integer
          nullable: true
          description: Line number in the old file (null for additions)
        new_line_num:
          type: integer
          nullable: true
          description: Line number in the new file (null for deletions)
    DiffHunk:
      type: object
      required: [header, old_start, new_start, lines]
      properties:
        header:
          type: string
          description: The @@ ... @@ hunk header line
        old_start:
          type: integer
          description: Starting line number in the old file
        new_start:
          type: integer
          description: Starting line number in the new file
        lines:
          type: array
          items:
            $ref: '#/components/schemas/DiffLine'
    DiffFile:
      type: object
      required: [path, change_type, additions, deletions, binary, hunks]
      properties:
        path:
          type: string
          description: File path (new path for renamed files)
        old_path:
          type: string
          nullable: true
          description: Original file path (only set for renamed files)
        change_type:
          type: string
          enum: [added, modified, deleted, renamed]
        additions:
          type: integer
          description: Number of added lines
        deletions:
          type: integer
          description: Number of deleted lines
        binary:
          type: boolean
          description: True if this is a binary file
        hunks:
          type: array
          items:
            $ref: '#/components/schemas/DiffHunk'
    DiffResponse:
      type: object
      required: [files, base_ref, head_ref]
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/DiffFile'
        base_ref:
          type: string
          description: The base ref used for this diff
        head_ref:
          type: string
          description: The head ref used for this diff
        merge_conflict:
          type: boolean
          description: True if there are merge conflicts between base_ref and head_ref
        conflict_files:
          type: array
          description: List of files with merge conflicts (populated when merge_conflict is true)
          items:
            type: string
        uncommitted_changes:
          type: boolean
          description: True if there are uncommitted changes in the worktree
        uncommitted_summary:
          description: Counts of uncommitted changes in the worktree
          $ref: '#/components/schemas/UncommittedSummary'
        behind_count:
          type: integer
          description: Number of commits on the base branch not yet merged into the agent branch (how far behind the base the branch is)
        base_commit:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CommitInfo'
          description: Details of the base commit (if a specific commit SHA was given)
        head_commit:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CommitInfo'
          description: Details of the head commit
    UncommittedSummary:
      type: object
      required: [tracked_count, untracked_count]
      properties:
        tracked_count:
          type: integer
          description: Number of tracked files with staged or unstaged changes
        untracked_count:
          type: integer
          description: Number of untracked (new, never-added) files
...
    ArtifactFile:
      type: object
      required: [name, change_type]
      properties:
        name:
          type: string
          description: File name relative to the artifact set
        change_type:
          type: string
          enum: [added, removed, modified, unchanged]
        left_url:
          type: string
          nullable: true
          description: URL of the file for the left version (null if absent on the left)
        right_url:
          type: string
          nullable: true
          description: URL of the file for the right version (null if absent on the right)
        tags:
          type: array
          nullable: true
          items:
            type: string
          description: >-
            Labels for this file, read from a sibling JSON sidecar
            (<file>.meta, {"tags": [...]}). A "category::value" tag is a
            scoped label — only one value per category survives. Drives the
            artifacts panel's tag badges and filter. Null/absent when the file
            has no tags.
        unverified:
          type: boolean
          nullable: true
          description: >-
            True only for a video file reported as "modified" whose verdict is a
            raw byte-hash comparison because ffmpeg was unavailable to verify it
            frame-by-frame — so the change may be spurious (e.g. only container
            metadata differs). Absent/false for images and for frame-verified
            video. The UI shows a caveat badge when set.
        fps:
          type: number
          format: double
          nullable: true
          description: >-
            Frame rate of a video file, read from its sibling JSON sidecar
            (<file>.meta, {"fps": 60}). HTML5 video exposes no frame rate, so the
            viewer's frame-step buttons use it to size a single-frame step.
            Null/absent when the sidecar omits it, in which case the viewer
            assumes a sensible default. Only meaningful for video files.
    ArtifactSet:
      type: object
      required: [name, status, changed, files]
      properties:
        name:
          type: string
          description: The configured artifact script name
        status:
          type: string
          enum: [ready, generating, error]
        changed:
          type: boolean
          description: Whether any file differs between the two versions
        error:
          type: string
          nullable: true
          description: >-
            Set only when the whole set failed (status "error") — i.e. both
            sides failed, or a side could not be loaded at all. When just one
            side fails while the other renders, status stays "ready" and the
            failure is reported in left_error / right_error instead.
        left_error:
          type: string
          nullable: true
          description: >-
            Error message from the LEFT (before) generation when that side
            failed but the RIGHT side still rendered, so the panel can show the
            available images alongside a warning. Null when the left side
            succeeded (or when the whole set failed — see error).
        right_error:
          type: string
          nullable: true
          description: As left_error, for the RIGHT (after) side.
        started_at:
          type: integer
          format: int64
          nullable: true
          description: >-
            Unix time (seconds) the earliest in-flight side started, so the UI
            can show how long it has been running. Only set while status is
            "generating".
        left_progress:
          type: string
          nullable: true
          description: >-
            Latest progress line of the in-flight LEFT (before) generation. Taken
            from `::hydra:progress::` marker lines the script emits, falling back
            to the latest stdout line until the first marker is seen. Only set
            while that side is generating.
        right_progress:
          type: string
          nullable: true
          description: As left_progress, for the RIGHT (after) generation.
        left_log:
          type: array
          nullable: true
          description: >-
            Captured stdout+stderr lines of the in-flight LEFT (before)
            generation, surfaced as a live log. Only populated while that side is
            generating; once settled, fetch left_log_url instead.
          items:
            $ref: '#/components/schemas/ArtifactLogLine'
        right_log:
          type: array
          nullable: true
          description: As left_log, for the RIGHT (after) generation.
          items:
            $ref: '#/components/schemas/ArtifactLogLine'
        left_log_url:
          type: string
          nullable: true
          description: >-
            URL to fetch the persisted build log of the LEFT (before) side once
            it has settled (ready or error), so the log can be reopened after
            generation finishes. Null while generating or if no log was captured.
        right_log_url:
          type: string
          nullable: true
          description: As left_log_url, for the RIGHT (after) side.
        pending_tags:
          type: array
          nullable: true
          description: >-
            Tags already known from a side that has settled while the other side
            is still generating, so the diff viewer's tag filter can appear
            before the whole set is ready (the "before" side often finishes
            first thanks to caching). Deduped and sorted, drawn from whichever
            side(s) have produced files so far. Only set while status is
            "generating"; once ready, tags are carried per file in files[].tags.
          items:
            type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactFile'
    ArtifactLogLine:
      type: object
      required: [text, stream]
      properties:
        text:
          type: string
          description: One captured output line (no trailing newline)
        stream:
          type: string
          enum: [stdout, stderr]
          description: Which stream the line came from; stderr is rendered in red
    ArtifactsResponse:
      type: object
      required: [scripts]
      properties:
        scripts:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactSet'
...